+++ neoforge/src/main/java/org/thinkingstudio/ryoamiclights/neoforge/mixin/VideoOptionsScreenMixin.java
@@ -8,7 +8,7 @@
  * see the LICENSE file.
  */
 
-package org.thinkingstudio.ryoamiclights.fabric.mixin;
+package org.thinkingstudio.ryoamiclights.neoforge.mixin;
 
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.screen.option.GameOptionsScreen;