+++ fabric/CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
-- remove entrypoint
+- update ObsidianUI, FabricAPI and NeoForge version
+- (NeoForge) port `VideoOptionsScreenMixin` from Fabric side